feat: add webpack and handler of internal server error
