+++ src/validator.js
