+++ webpack.config.js
@@ -1,0 +1,12 @@
+const path = require("path");
+
+module.exports = {
+  watch: true,
+  target: "node",
+  entry: "./src/server.js",
+  output: {
+    filename: "bundle.js",
+    path: path.resolve(__dirname, "build"),
+    publicPath: "/",
+  },
+};
